refactor: 宣言が重複している型を削除(#19)
