+++ chrome/src/utils/history.ts
@@ -1,16 +1,5 @@
 import { KEYS } from '@/enums';
-
-export type Prices = number[];
-export type History = Readonly<{
-  id: string;
-  title: string;
-  href: string;
-  imageUrl: string;
-  prices: Prices;
-  salePrices: Prices | null;
-  saleLimitTime: string | null;
-}>;
-type Histories = History[] | [];
+import { History, Histories } from '@/models/history';
 
 export function add(histories: Histories, newHistory: History): Histories {
   return [newHistory, ...histories.filter(({ id }) => id !== newHistory.id)];
